# docs/swagger.yaml
basePath: /api/v1
definitions:
  achievement.CreateAchievementRequest:
    properties:
      category:
        type: string
      dateAchieved:
        type: string
      projectID:
        type: string
      title:
        type: string
    required:
    - category
    - dateAchieved
    - title
    type: object
  achievement.UpdateAchievementRequest:
    properties:
      category:
        type: string
      dateAchieved:
        type: string
      projectID:
        type: string
      title:
        type: string
    type: object
  models.Achievement:
    properties:
      category:
        type: string
      dateAchieved:
        type: string
      id:
        type: string
      projectID:
        type: string
      title:
        type: string
      userID:
        type: string
    type: object
  models.Experience:
    properties:
      company:
        type: string
      description:
        type: string
      endDate:
        description: ใช้ pointer เพื่อให้เป็น NULL ได้กรณีปัจจุบันยังทำอยู่
        type: string
      id:
        type: string
      jobTitle:
        type: string
      projects:
        items:
          $ref: '#/definitions/models.Project'
        type: array
      startDate:
        type: string
      userID:
        type: string
    type: object
  models.Project:
    properties:
      coverImageURL:
        type: string
      createdAt:
        type: string
      description:
        type: string
      experience:
        $ref: '#/definitions/models.Experience'
      experienceID:
        type: string
      githubURL:
        type: string
      id:
        type: string
      images:
        items:
          $ref: '#/definitions/models.ProjectImage'
        type: array
      technologies:
        items:
          $ref: '#/definitions/models.Technology'
        type: array
      title:
        type: string
      userID:
        type: string
    type: object
  models.ProjectImage:
    properties:
      caption:
        type: string
      id:
        type: string
      imageURL:
        type: string
      projectID:
        type: string
    type: object
  models.Study:
    properties:
      degree:
        type: string
      faculty:
        type: string
      gpa:
        type: number
      graduationDate:
        type: string
      id:
        type: string
      institution:
        type: string
      major:
        type: string
      userID:
        type: string
    type: object
  models.Technology:
    properties:
      category:
        type: string
      iconURL:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
info:
  contact:
    name: Piyadanai Krongklang
    url: https://github.com/Piyadanai03
  description: นี่คือระบบ Backend สำหรับ Portfolio ของคุณปิยดนัย
  termsOfService: http://swagger.io/terms/
  title: Piyadanai Portfolio API
  version: "1.0"
paths:
  /educations:
    get:
      description: ดึงข้อมูลการศึกษาทั้งหมด
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Study'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: ดึงรายการการศึกษา
      tags:
      - Education
  /educations/{id}:
    get:
      description: ดึงข้อมูลการศึกษาตาม ID
      parameters:
      - description: ID ของการศึกษา
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Study'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: ดูข้อมูลการศึกษาตาม ID
      tags:
      - Education
  /login:
    post:
      consumes:
      - application/json
      description: ตรวจสอบ Username/Password และคืนค่า JWT Token
      parameters:
      - description: ข้อมูล Login (username, password)
        in: body
        name: input
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: เข้าสู่ระบบ (Admin)
      tags:
      - Authentication
  /member/achievement:
    get:
      consumes:
      - application/json
      description: ดึงข้อมูลความสำเร็จทั้งหมดจากฐานข้อมูล (เรียงจากใหม่ไปเก่า)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Achievement'
            type: array
      security:
      - BearerAuth: []
      summary: ดึงข้อมูลความสำเร็จทั้งหมด
      tags:
      - Achievement
    post:
      consumes:
      - application/json
      description: สร้างข้อมูลความสำเร็จใหม่ (ต้อง Login)
      parameters:
      - description: Achievement data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/achievement.CreateAchievementRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: สร้างข้อมูลความสำเร็จ
      tags:
      - Achievement
  /member/achievement/{id}:
    delete:
      consumes:
      - application/json
      description: ลบข้อมูลความสำเร็จออกจากระบบตาม ID (ต้อง Login)
      parameters:
      - description: ID ของความสำเร็จ
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: ลบข้อมูลความสำเร็จ
      tags:
      - Achievement
    get:
      consumes:
      - application/json
      description: ดึงข้อมูลความสำเร็จตาม ID จากฐานข้อมูล
      parameters:
      - description: Achievement ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Achievement'
      security:
      - BearerAuth: []
      summary: ดึงข้อมูลความสำเร็จตาม ID
      tags:
      - Achievement
    put:
      consumes:
      - application/json
      description: อัพเดตข้อมูลความสำเร็จ (ต้อง Login)
      parameters:
      - description: ID ของความสำเร็จ
        in: path
        name: id
        required: true
        type: string
      - description: Achievement data to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/achievement.UpdateAchievementRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: แก้ไขข้อมูลความสำเร็จ
      tags:
      - Achievement
  /member/education:
    post:
      consumes:
      - application/json
      description: บันทึกข้อมูลการศึกษาใหม่ (ต้อง Login)
      parameters:
      - description: ข้อมูลการศึกษา
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Study'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: เพิ่มข้อมูลการศึกษา
      tags:
      - Education
  /member/education/{id}:
    delete:
      description: ลบข้อมูลการศึกษาตาม ID (ต้อง Login)
      parameters:
      - description: ID ของการศึกษา
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: ลบข้อมูลการศึกษา
      tags:
      - Education
    put:
      consumes:
      - application/json
      description: อัพเดตข้อมูลการศึกษาตาม ID (ต้อง Login)
      parameters:
      - description: ID ของการศึกษา
        in: path
        name: id
        required: true
        type: string
      - description: ข้อมูลการศึกษาที่ต้องการแก้ไข
        in: body
        name: input
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: แก้ไขข้อมูลการศึกษา
      tags:
      - Education
  /member/experience:
    post:
      consumes:
      - application/json
      description: บันทึกข้อมูลประสบการณ์ทำงานใหม่ (ต้อง Login)
      parameters:
      - description: ข้อมูลประสบการณ์ทำงาน
        in: body
        name: input
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: เพิ่มข้อมูลประสบการณ์ทำงาน
      tags:
      - Experience
  /member/experience/{id}:
    delete:
      description: ลบข้อมูลประสบการณ์ทำงานตาม ID (ต้อง Login)
      parameters:
      - description: ID ของประสบการณ์ทำงาน
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: ลบข้อมูลประสบการณ์ทำงาน
      tags:
      - Experience
    put:
      consumes:
      - application/json
      description: อัพเดตข้อมูลประสบการณ์ทำงาน (ต้อง Login)
      parameters:
      - description: ID ของประสบการณ์ทำงาน
        in: path
        name: id
        required: true
        type: string
      - description: ข้อมูลประสบการณ์ทำงานที่ต้องการแก้ไข
        in: body
        name: input
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: แก้ไขข้อมูลประสบการณ์ทำงาน
      tags:
      - Experience
  /member/experiences:
    get:
      description: ดึงข้อมูลประสบการณ์ทั้งหมด พร้อมโปรเจกต์ที่เกี่ยวข้อง
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Experience'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: ดึงรายการประสบการณ์ทำงาน
      tags:
      - Experience
  /member/experiences/{id}:
    get:
      description: ดึงข้อมูลประสบการณ์ตาม ID
      parameters:
      - description: ID ของประสบการณ์
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Experience'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: ดูประสบการณ์ทำงานตาม ID
      tags:
      - Experience
  /member/projects/{id}:
    delete:
      description: ลบข้อมูลโปรเจกต์ตาม ID (ต้อง Login)
      parameters:
      - description: ID ของโปรเจกต์
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: ลบข้อมูลโปรเจกต์
      tags:
      - Projects
  /member/tech:
    get:
      consumes:
      - application/json
      description: ดึงข้อมูลเทคโนโลยีทั้งหมดจากฐานข้อมูล
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Technology'
            type: array
      security:
      - BearerAuth: []
      summary: ดึงข้อมูลเทคโนโลยีทั้งหมด
      tags:
      - Technology
  /member/tech/{id}:
    delete:
      consumes:
      - application/json
      description: ลบข้อมูลเทคโนโลยีออกจากระบบตาม ID (ต้อง Login)
      parameters:
      - description: ID ของเทคโนโลยี
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: ลบข้อมูลเทคโนโลยี
      tags:
      - Technology
    get:
      consumes:
      - application/json
      description: ดึงข้อมูลเทคโนโลยีจากฐานข้อมูลตาม ID ที่ระบุ
      parameters:
      - description: ID ของเทคโนโลยี
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Technology'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: ดึงข้อมูลเทคโนโลยีตาม ID
      tags:
      - Technology
    put:
      consumes:
      - multipart/form-data
      description: อัพเดตข้อมูลเทคโนโลยี (เช่น ชื่อ, หมวดหมู่, และเปลี่ยนรูปไอคอน)
        (ต้อง Login)
      parameters:
      - description: ID ของเทคโนโลยี
        in: path
        name: id
        required: true
        type: string
      - description: ชื่อเทคโนโลยี
        in: formData
        name: name
        type: string
      - description: หมวดหมู่เทคโนโลยี
        in: formData
        name: category
        type: string
      - description: ไฟล์รูปภาพไอคอน
        in: formData
        name: icon
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: แก้ไขข้อมูลเทคโนโลยี
      tags:
      - Technology
securityDefinitions:
  BearerAuth:
    description: ใส่คำว่า "Bearer " ตามด้วย Token ของคุณ
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"